refactor(#155 PR 6/10): extract ui/campaign-ui.js and persistence/save-load.js
Pure code movement, zero behavior change. game.js: 2969 -> 2318 lines
(4076 at the start of the split).

- src/ui/campaign-ui.js (353 lines): the 15 campaign UI functions —
  level select, briefing, debrief, objectives panel, progress label,
  toolbar gating, level tooltips, retry/next flow. window-assigned
  arrows became named exports; thin window.x = importedX assignments
  stay in game.js's boundary block.
- src/persistence/save-load.js (411 lines): the full save/load system —
  save modal, saveGameState, downloadSaveFile, upload, migrateOldSave,
  loadGameState, clearCurrentGame, restoreServices/Connections.
- campaign.js now imports its UI hooks from ui/campaign-ui.js and no
  longer imports game.js at all.
- tests/i18n-usage scans src/ui and src/persistence too.

Verified: lint 0, 87/87 tests, 28-module graph links, in-browser —
campaign select renders, level 2 starts with correct budget and toolbar
gating (db locked, s3 allowed), objectives panel updates through the
repointed import, 40/40 requests processed, save -> continue restores
money, save-file upload round-trip clean, zero console errors.
